.env, авторизация, доработка апи

+++ src/controllers/order.controller.ts
@@ -13,8 +13,10 @@ const newOrder = async (req: Request, res: Response) => {
    }
     */
     try {
-        const user = await getUser(req.body.userId)
-        if (user) {
+        const user = await getUser(req.body.userId);
+        if (!req.user) {
+            res.status(401).send('Unauthorized');
+        } else if (user) {
             createOrder(req.body.userId, req.body.amount).then(data => res.status(201).send(data)).catch(err => {
                 console.error(err.message + '\n' + err.stack)
                 res.status(502).send(null)
@@ -29,7 +31,9 @@ const getOrderById = (req: Request, res: Response) => {
     /* 	#swagger.tags = ['Orders']
     #swagger.description = 'get order' */
     try {
-        if (isNaN(req.params.id)) {
+        if (!req.user) {
+            res.status(401).send('Unauthorized');
+        } else if (isNaN(req.params.id)) {
             res.status(400).send('Bad request');
         } else {
             getOrder(req.params.id).then(data => {
@@ -52,16 +56,22 @@ const updateOrderById = (req: Request, res: Response) => {
     /* 	#swagger.tags = ['Orders']
     #swagger.description = 'update order' */
     try {
-        updateOrder(req.params.id, req.body).then(data => {
-            if (data) {
-                res.status(200).send(data);
-            } else {
-                res.status(404).send('Not Found');
-            }
-        }).catch(err => {
-            console.error(err.message + '\n' + err.stack)
-            res.status(502).send(null)
-        })
+        if (!req.user) {
+            res.status(401).send('Unauthorized');
+        } else if (isNaN(req.params.id)) {
+            res.status(400).send('Bad request');
+        } else {
+            updateOrder(req.params.id, req.body).then(data => {
+                if (data) {
+                    res.status(200).send(data);
+                } else {
+                    res.status(404).send('Not Found');
+                }
+            }).catch(err => {
+                console.error(err.message + '\n' + err.stack)
+                res.status(502).send(null)
+            })
+        }
     } catch (e) {
         console.error(e);
     }
@@ -71,16 +81,22 @@ const deleteOrderById = (req: Request, res: Response) => {
     /* 	#swagger.tags = ['Orders']
     #swagger.description = 'delete order' */
     try {
-        deleteOrder(req.params.id).then(data => {
-            if (data) {
-                res.status(200).send(data);
-            } else {
-                res.status(404).send('Not Found');
-            }
-        }).catch(err => {
-            console.error(err.message + '\n' + err.stack)
-            res.status(502).send(null)
-        })
+        if (!req.user) {
+            res.status(401).send('Unauthorized');
+        } else if (isNaN(req.params.id)) {
+            res.status(400).send('Bad request');
+        } else {
+            deleteOrder(req.params.id).then(data => {
+                if (data) {
+                    res.status(200).send(data);
+                } else {
+                    res.status(404).send('Not Found');
+                }
+            }).catch(err => {
+                console.error(err.message + '\n' + err.stack)
+                res.status(502).send(null)
+            })
+        }
     } catch (e) {
         console.error(e);
     }
@@ -88,12 +104,19 @@ const deleteOrderById = (req: Request, res: Response) => {
 
 const getOrdersWithFilter = (req: Request, res: Response) => {
     /* 	#swagger.tags = ['Orders']
-    #swagger.description = 'get orders' */
+    #swagger.description = 'get orders'
+    #swagger.security = [{
+            "Bearer": []
+    }] */
     try {
-        getOrders().then(data => res.status(200).send(data)).catch(err => {
-            console.error(err.message + '\n' + err.stack)
-            res.status(502).send(null)
-        })
+        if (!req.user) {
+            res.status(401).send('Unauthorized');
+        } else {
+            getOrders().then(data => res.status(200).send(data)).catch(err => {
+                console.error(err.message + '\n' + err.stack)
+                res.status(502).send(null)
+            })
+        }
     } catch (e) {
         console.error(e);
     }
